fix sbw sdk v4 (#770)
* refactor: Add missing entrypointAddress parameter to getSmartWallet

This commit adds the missing entrypointAddress parameter to the getSmartWallet function in the getSmartWallet.ts file. The entrypointAddress is now passed as an argument to the SmartWallet constructor. This change ensures that the entrypointAddress is correctly set when using the v4 SDK with smart backend wallets.

* fix: Set accountFactoryAddress and entrypointAddress in getWallet

This commit updates the getWallet function in the getWallet.ts file to set the accountFactoryAddress and entrypointAddress parameters when creating a smart wallet. The accountFactoryAddress and entrypointAddress are now retrieved from the walletDetails object and passed as arguments to the SmartWallet constructor. This change ensures that the accountFactoryAddress and entrypointAddress are correctly set when using the v4 SDK with smart backend wallets.

* fix: Set accountFactoryAddress and entrypointAddress in insertTransaction

This commit updates the insertTransaction function in the insertTransaction.ts file to set the accountFactoryAddress and entrypointAddress parameters when creating a smart wallet. The accountFactoryAddress and entrypointAddress are now retrieved from the walletDetails object and assigned to the queuedTransaction object. This change ensures that the accountFactoryAddress and entrypointAddress are correctly set when using the v4 SDK with smart backend wallets.

* chore: Clean up smart-local-wallet.test.ts

This commit removes unused imports and updates the chainId variable in the smart-local-wallet.test.ts file. The chainId is now retrieved from the chain object. This change improves the readability and maintainability of the code.

* feat: Add smart-local-wallet-sdk-v4.test.ts

This commit adds the smart-local-wallet-sdk-v4.test.ts file, which contains tests for creating a local smart backend wallet and sending SDK v4 transactions. The tests deploy an ERC20 token, mint tokens, and check the balance. This change improves the test coverage for the smart backend wallet functionality.